# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  docs.ChannelCreateRequest:
    properties:
      base_url:
        example: https://api.openai.com/v1
        type: string
      config:
        example: '{}'
        type: string
      id:
        example: openai-main
        type: string
      key:
        example: sk-***
        type: string
      model_configs:
        items:
          $ref: '#/definitions/docs.ChannelModelConfigRequest'
        type: array
      models:
        example: gpt-4o-mini,gpt-4o
        type: string
      name:
        example: OpenAI Main
        type: string
      priority:
        example: 0
        type: integer
      protocol:
        example: openai
        type: string
      status:
        example: 1
        type: integer
      system_prompt:
        example: ""
        type: string
      test_model:
        example: gpt-4o-mini
        type: string
      weight:
        example: 0
        type: integer
    type: object
  docs.ChannelModelConfigRequest:
    properties:
      currency:
        example: USD
        type: string
      input_price:
        example: 1e-06
        type: number
      model:
        example: gpt-4o-mini
        type: string
      output_price:
        example: 2e-06
        type: number
      price_unit:
        example: usd_per_token
        type: string
      selected:
        example: true
        type: boolean
      sort_order:
        example: 1
        type: integer
      upstream_model:
        example: gpt-4o-mini
        type: string
    type: object
...
  docs.ChannelUpdateRequest:
    properties:
      base_url:
        example: https://api.openai.com/v1
        type: string
      config:
        example: '{}'
        type: string
      id:
        example: openai-main
        type: string
      key:
        example: sk-***
        type: string
      model_configs:
        items:
          $ref: '#/definitions/docs.ChannelModelConfigRequest'
        type: array
      models:
        example: gpt-4o-mini,gpt-4o
        type: string
      name:
        example: OpenAI Main
        type: string
      priority:
        example: 0
        type: integer
      protocol:
        example: openai
        type: string
      status:
        example: 1
        type: integer
      system_prompt:
        example: ""
        type: string
      test_model:
        example: gpt-4o-mini
        type: string
      weight:
        example: 0
        type: integer
    type: object
  docs.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: integer
    type: object
...
  docs.StandardResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: integer
    type: object
...
host: router.yeying.pub
info:
  contact: {}
  description: Router public/admin API (JWT/UCAN) under /api/v1.
  title: Router API
  version: v1
paths:
  /api/v1/admin/channel:
    get:
      parameters:
      - description: Page index
        in: query
        name: p
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List channels (admin)
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: Channel payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/docs.ChannelCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create channel (admin)
      tags:
      - admin
    put:
      consumes:
      - application/json
      parameters:
      - description: Channel update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/docs.ChannelUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update channel (admin)
      tags:
      - admin
  /api/v1/admin/channel/{id}:
    delete:
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete channel (admin)
      tags:
      - admin
    get:
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get channel by ID (admin)
      tags:
      - admin
  /api/v1/admin/channel/disabled:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete disabled channels (admin)
      tags:
      - admin
  /api/v1/admin/channel/draft:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create channel draft (admin)
      tags:
      - admin
...
  /api/v1/admin/channel/search:
    get:
      parameters:
      - description: Keyword
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search channels (admin)
      tags:
      - admin
...
  /api/v1/admin/channel/test_model:
    put:
      consumes:
      - application/json
      parameters:
      - description: Channel test model payload
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update channel test model (admin)
      tags:
      - admin
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"